+++ CHANGES.rst
@@ -1,5 +1,19 @@
 CHANGE LOG
 ==========
+
+0.2.7
+-----
+- Fixed bug in slack stats integration
+
+0.2.6
+-----
+- Added the option of exporting user data as CSV in django admin
+
+0.2.5
+-----
+- Added cellphone number to registration
+- Added User Profiles Settings in wagtail
+
 0.2.4
 -----
 - Removed requirement for date of birth when editing profile